protections

+++ srcs/string/ft_strcmpi.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strcmp.c                                        :+:      :+:    :+:   */
+/*   ft_strcmpi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kdumarai <kdumarai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 19:11:27 by kdumarai          #+#    #+#             */
-/*   Updated: 2018/01/30 21:48:44 by kdumarai         ###   ########.fr       */
+/*   Updated: 2018/07/28 17:46:30 by kdumarai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strcmpi(const char *s1, const char *s2)
+unsigned long	ft_strcmpi(const char *s1, const char *s2)
 {
-	unsigned int	i;
+	unsigned long	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && s1[i] == s2[i])
